# Conflicts:
#	src/main/java/com/sep490/g28/hvh/be/constant/ENotificationType.java
#	src/main/java/com/sep490/g28/hvh/be/exception/errorCodeImpl/EventErrorCode.java
#	src/main/java/com/sep490/g28/hvh/be/exception/errorCodeImpl/RateAndReviewErrorCode.java
#	src/main/java/com/sep490/g28/hvh/be/exception/errorCodeImpl/ValidationErrorCode.java
#	src/main/java/com/sep490/g28/hvh/be/repository/EventApplicationRepository.java
#	src/main/java/com/sep490/g28/hvh/be/service/impl/EventApplicationServiceImpl.java
#	src/test/java/com/sep490/g28/hvh/be/service/EventApplicationServiceTest.java

+++ src/test/java/com/sep490/g28/hvh/be/service/EventApplicationServiceTest.java
@@ -137,8 +137,8 @@ public class EventApplicationServiceTest {
         Volunteer vol = new Volunteer();
         vol.setId(UUID.randomUUID());
         vol.setAddress("Hanoi");
-        vol.setHonorScore((short) 10);
-        vol.setCreditScore((short) 10);
+        vol.setHonorScore(10);
+        vol.setCreditScore(10);
         vol.setCreatedAt(OffsetDateTime.now());
 
         EventApplication app = new EventApplication();
@@ -573,7 +573,7 @@ public class EventApplicationServiceTest {
 
         service.cancelApplication(app.getId());
 
-        assertEquals((short) 7, app.getVolunteer().getHonorScore());
+        assertEquals(7, app.getVolunteer().getHonorScore());
 
         verify(volunteerRepository).save(app.getVolunteer());
         verify(eventSessionRepository).save(app.getSession());
@@ -605,7 +605,7 @@ public class EventApplicationServiceTest {
 
         service.cancelApplication(app.getId());
 
-        assertEquals((short) 10, app.getVolunteer().getHonorScore());
+        assertEquals(10, app.getVolunteer().getHonorScore());
 
         verify(eventSessionRepository).save(app.getSession());
 
@@ -697,7 +697,7 @@ public class EventApplicationServiceTest {
 
         service.cancelApplication(app.getId());
 
-        assertEquals((short) 7, app.getVolunteer().getHonorScore());
+        assertEquals(7, app.getVolunteer().getHonorScore());
 
         verify(volunteerRepository).save(app.getVolunteer());
 
@@ -753,7 +753,7 @@ public class EventApplicationServiceTest {
 
         service.cancelApplication(app.getId());
 
-        assertEquals((short)10, app.getVolunteer().getHonorScore());
+        assertEquals(10, app.getVolunteer().getHonorScore());
 
         verify(volunteerRepository, never()).save(any());
     }
